electronics_network/requirements.txt
django==4.2
asgiref==3.8.1
tzdata==2024.1
sqlparse==0.4.4
djangorestframework==3.14.0
psycopg2-binary==2.9.9
pillow==10.2.0
python-dotenv==1.0.0
django-filter==24.2
djangorestframework-simplejwt==5.3.0
PyJWT==2.8.0
drf-yasg==1.21.7